+++ src/index.html
@@ -4,16 +4,14 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Weather App</title>
-    <link rel="stylesheet" href="styles.css" />
   </head>
   <body>
     <header>
-      * Click and edit city name to view different weather forecasts.
+      * Click and edit city name below to view different weather forecasts.
     </header>
     <main></main>
     <footer>
       <a href="https://github.com/xtang999" target="_blank">&copy; xtang</a>
     </footer>
-    <script src="script.js"></script>
   </body>
 </html>
