+++ src/index.js
@@ -1,0 +1,6 @@
+// index.js
+
+import './styles/styles.css';
+import { init } from './app.js';
+
+init();
